requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
SQLAlchemy==2.0.25
pyhumps==3.8.0
